Split into server+client+serverclient.
